[Refactor] Organize file structure
