+++ src/index.jsx
@@ -3,7 +3,7 @@ import ReactDOM from 'react-dom';
 
 import { Provider } from 'react-redux';
 
-import store from './store';
+import store from './redux/store';
 
 import App from './App';
