+++ src/regions/Regions.test.jsx
@@ -4,7 +4,7 @@ import { render } from '@testing-library/react';
 
 import Regions from './Regions';
 
-import regions from '../fixtures/regions';
+import regions from '../../fixtures/regions';
 
 test('Regions', () => {
   const { queryByText } = render((
